+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugs fixed
 
 ### Enhancements
+- GLS-58 - Updated CPI data source csv
 
 ## [2.2.2](https://github.com/uktrade/directory-api/releases/tag/2.2.2)
 [Full Changelog](https://github.com/uktrade/directory-api/compare/2.2.1...2.2.2)